Split out subclass.
