+++ src/app/loadout/Loadouts.m.scss
@@ -96,22 +96,6 @@
   margin-right: 8px;
 }
 
-.subclassContainer {
-  --item-icon-size: calc(0.8 * var(--item-size));
-  composes: flexRow from '../dim-ui/common.m.scss';
-  width: calc(var(--item-icon-size) + 3 * var(--item-icon-size) + 2 * var(--item-margin));
-}
-
-.subclass {
-  width: var(--item-size);
-  min-width: var(--item-size);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-right: var(--item-margin);
-  @include vertical-space-children(var(--item-margin));
-}
-
 .itemCategory {
   display: flex;
   flex-direction: column;
@@ -177,18 +161,6 @@
   }
 }
 
-.power {
-  display: inline;
-  color: $gold;
-  flex: inherit;
-  font-size: 12px;
-  letter-spacing: normal;
-  svg {
-    margin-right: 2px;
-    font-size: 10px;
-  }
-}
-
 .mods {
   flex: 1;
 }
@@ -208,11 +180,6 @@
 
 .showModPlacementButton {
   margin-top: 8px;
-}
-
-.subclassMods {
-  composes: modsGrid;
-  min-width: calc(3 * var(--item-icon-size) + 2 * var(--item-margin));
 }
 
 .modsPlaceholder {
