+++ src/app/loadout/Loadouts.tsx
@@ -1,6 +1,5 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import { DestinyAccount } from 'app/accounts/destiny-account';
-import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { languageSelector } from 'app/dim-api/selectors';
 import BungieImage from 'app/dim-ui/BungieImage';
 import CharacterSelect from 'app/dim-ui/CharacterSelect';
@@ -14,7 +13,6 @@ import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import { allItemsSelector, bucketsSelector, sortedStoresSelector } from 'app/inventory/selectors';
 import { DimStore } from 'app/inventory/store-types';
 import { useLoadStores } from 'app/inventory/store/hooks';
-import { isPluggableItem } from 'app/inventory/store/sockets';
 import { getCurrentStore, getStore } from 'app/inventory/stores-helpers';
 import { SelectedArmorUpgrade } from 'app/loadout-builder/filter/ArmorUpgradePicker';
 import ExoticArmorChoice from 'app/loadout-builder/filter/ExoticArmorChoice';
@@ -36,29 +34,22 @@ import { showNotification } from 'app/notifications/notifications';
 import { startWordRegexp } from 'app/search/search-filters/freeform';
 import { useSetting } from 'app/settings/hooks';
 import { LoadoutSort } from 'app/settings/initial-settings';
-import {
-  addIcon,
-  AppIcon,
-  faCalculator,
-  faExclamationTriangle,
-  powerActionIcon,
-  searchIcon,
-} from 'app/shell/icons';
+import { addIcon, AppIcon, faCalculator, faExclamationTriangle, searchIcon } from 'app/shell/icons';
 import { querySelector, useIsPhonePortrait } from 'app/shell/selectors';
 import { LoadoutStats } from 'app/store-stats/CharacterStats';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { itemCanBeEquippedBy } from 'app/utils/item-utils';
-import { getSocketsByIndexes } from 'app/utils/socket-utils';
 import { copyString } from 'app/utils/util';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
-import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
+import { BucketHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
 import React, { useMemo, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import PlugDef from './loadout-ui/PlugDef';
+import { Subclass } from './loadout-ui/Subclass';
 import styles from './Loadouts.m.scss';
 import ModAssignmentDrawer from './mod-assignment-drawer/ModAssignmentDrawer';
 import { createGetModRenderKey } from './mod-utils';
@@ -351,80 +342,6 @@ function LoadoutRow({
   );
 }
 
-function Subclass({
-  defs,
-  subclass,
-  power,
-}: {
-  defs: D2ManifestDefinitions;
-  subclass?: DimLoadoutItem;
-  power: number;
-}) {
-  const getModRenderKey = createGetModRenderKey();
-  const plugs = useMemo(() => {
-    const plugs: PluggableInventoryItemDefinition[] = [];
-
-    if (subclass?.sockets?.categories) {
-      for (const category of subclass.sockets.categories) {
-        const showInitial =
-          category.category.hash !== SocketCategoryHashes.Aspects &&
-          category.category.hash !== SocketCategoryHashes.Fragments;
-        const sockets = getSocketsByIndexes(subclass.sockets, category.socketIndexes);
-
-        for (const socket of sockets) {
-          const override = subclass.socketOverrides?.[socket.socketIndex];
-          const initial = socket.socketDefinition.singleInitialItemHash;
-          const hash = override || (showInitial && initial);
-          const plug = hash && defs.InventoryItem.get(hash);
-          if (plug && isPluggableItem(plug)) {
-            plugs.push(plug);
-          }
-        }
-      }
-    }
-
-    return plugs;
-  }, [subclass, defs]);
-
-  return (
-    <div className={styles.subclassContainer}>
-      <div className={styles.subclass}>
-        {subclass ? (
-          <ItemPopupTrigger item={subclass}>
-            {(ref, onClick) => (
-              <ConnectedInventoryItem
-                innerRef={ref}
-                // Disable the popup when plugs are available as we are showing
-                // plugs in the loadout and they may be different to the popup
-                onClick={plugs.length ? undefined : onClick}
-                item={subclass}
-                ignoreSelectedPerks
-              />
-            )}
-          </ItemPopupTrigger>
-        ) : (
-          <EmptyClassItem />
-        )}
-        {power !== 0 && (
-          <div className={styles.power}>
-            <AppIcon icon={powerActionIcon} />
-            <span>{power}</span>
-          </div>
-        )}
-      </div>
-      {plugs.length ? (
-        <div className={styles.subclassMods}>
-          {plugs?.map((plug) => (
-            <PlugDef key={getModRenderKey(plug)} plug={plug} />
-          ))}
-        </div>
-      ) : (
-        <div className={styles.modsPlaceholder}>{t('Loadouts.Abilities')}</div>
-      )}
-    </div>
-  );
-}
-
 function ItemCategory({
   category,
   subclass,
@@ -608,22 +525,5 @@ function LoadoutParametersDisplay({ params }: { params: LoadoutParameters }) {
         </div>
       )}
     </div>
-  );
-}
-
-function EmptyClassItem() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48">
-      <rect
-        transform="rotate(-45)"
-        y="17.470564"
-        x="-16.470564"
-        height="32.941124"
-        width="32.941124"
-        fill="rgba(255, 255, 255, 0.05)"
-        strokeWidth="1"
-        strokeMiterlimit="4"
-      />
-    </svg>
   );
 }
