+++ src/app/loadout/loadout-ui/Subclass.m.scss.d.ts
@@ -1,0 +1,11 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'modsPlaceholder': string;
+  'power': string;
+  'subclass': string;
+  'subclassContainer': string;
+  'subclassMods': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
